+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/PropagationPolicyConfigurable.java
@@ -13,13 +13,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.fabric8.kubernetes.client.dsl;
+package io.fabric8.kubernetes.client;
 
-import io.fabric8.kubernetes.client.GracePeriodConfigurable;
-import io.fabric8.kubernetes.client.PropagationPolicyConfigurable;
-
-public interface WatchListDeletable<T, L, B, H, W> extends VersionWatchable<H, W>, Listable<L>, Deletable<B>,
-                                                           GracePeriodConfigurable<Deletable<B>>,
-                                                           PropagationPolicyConfigurable<Deletable<B>>
+public interface PropagationPolicyConfigurable<T>
 {
+    T withPropagationPolicy(String propagationPolicy);
 }