+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/ListVisitFromServerGetDeleteRecreateWaitApplicable.java
@@ -21,10 +21,13 @@ import java.util.List;
 import io.fabric8.kubernetes.api.builder.Visitable;
 import io.fabric8.kubernetes.client.FromServerGettable;
 import io.fabric8.kubernetes.client.GracePeriodConfigurable;
+import io.fabric8.kubernetes.client.PropagationPolicyConfigurable;
 
 public interface ListVisitFromServerGetDeleteRecreateWaitApplicable<T, B> extends Visitable<ListVisitFromServerGetDeleteRecreateWaitApplicable<T, B>>,
                                                                           FromServerGettable<List<T>>, RecreateApplicable<List<T>, T>,
                                                                           CascadingDeletable<B>,
                                                                           Waitable<List<T>, T>,
-                                                                          GracePeriodConfigurable<CascadingDeletable<B>> {
+                                                                          GracePeriodConfigurable<CascadingDeletable<B>>,
+                                                                          PropagationPolicyConfigurable<CascadingDeletable<B>>
+{
 }