+++ kubernetes-client/src/main/java/io/fabric8/kubernetes/client/dsl/VisitFromServerGetWatchDeleteRecreateWaitApplicable.java
@@ -19,6 +19,7 @@ package io.fabric8.kubernetes.client.dsl;
 import io.fabric8.kubernetes.api.builder.Visitable;
 import io.fabric8.kubernetes.client.FromServerGettable;
 import io.fabric8.kubernetes.client.GracePeriodConfigurable;
+import io.fabric8.kubernetes.client.PropagationPolicyConfigurable;
 import io.fabric8.kubernetes.client.Watch;
 import io.fabric8.kubernetes.client.Watcher;
 
@@ -27,5 +28,6 @@ public interface VisitFromServerGetWatchDeleteRecreateWaitApplicable<T, B> exten
                                                                           CascadingDeletable<B>,
                                                                           Watchable<Watch, Watcher<T>>,
                                                                           Waitable<T, T>,
-                                                                          GracePeriodConfigurable<CascadingDeletable<B>> {
+                                                                          GracePeriodConfigurable<CascadingDeletable<B>>,
+                                                                          PropagationPolicyConfigurable<CascadingDeletable<B>> {
 }