+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## CHANGELOG
 
 ### 4.4-SNAPSHOT
+
 #### Bugs
   * Fix #1745: Calling getInputStreamFromDataOrFile function with correct parameter order
   * Fix #1730: Fix failing build on jdk11
@@ -8,6 +9,7 @@
   * Fixed Knative model so that it recognizes Container and Volume types as Buildable.
 
 #### Improvements
+  * Allow user to set a propagation policy on deletion
   * Added in kubernetes-server-mock (CRUD) the withoutLabel filter and respective tests
   * Removed @Valid annotation from all model class fields (improves quarkus integration).